requirements.txt
asgiref==3.2.5
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==3.0
django-allauth==0.41.0
django-cors-headers==3.2.1
django-filter==2.2.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
drf-yasg==1.17.1
idna==2.9
importlib-metadata==1.6.0
inflection==0.3.1
itypes==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.3
Pillow==7.0.0
psycopg2-binary==2.8.4
pyparsing==2.4.6
pyrsistent==0.16.0
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
sqlparse==0.3.1
swagger-spec-validator==2.5.0
uritemplate==3.0.1
urllib3==1.25.8
zipp==3.1.0
